refactor(material-consumptions): migrate components to feature directory and extract ConsumptionTable (Phase 6.1f)
- Move 'ConsumptionRow' from shared resources to 'features/material-consumptions/components'.
- Extract 'ConsumptionTable' component to handle list orchestration and filters.
- Update material consumptions list page to use new feature-specific components.
