+++ app/users/dashboard/resources/material-consumptions/page.tsx
@@ -2,28 +2,10 @@
 
 import { useState, useMemo } from 'react';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 import { routes } from '@/nav';
-import { Card, CardContent, CardHeader } from '@/components/shadcn/card';
+import { Card } from '@/components/shadcn/card';
 import { Button } from '@/components/shadcn/button';
-import { Badge } from '@/components/shadcn/badge';
-import { Input } from '@/components/shadcn/input';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/shadcn/select';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/shadcn/table';
-import { Pagination, PageHeader } from '@/components/common';
+import { PageHeader } from '@/components/common';
 import {
   Plus,
   Loader2,
@@ -31,31 +13,12 @@ import {
   CalendarDays,
   Package,
   BarChart3,
-  Search,
 } from 'lucide-react';
-import {
-  Empty,
-  EmptyMedia,
-  EmptyHeader,
-  EmptyTitle,
-  EmptyDescription,
-} from '@/components/shadcn/empty';
-import { format } from 'date-fns';
 import { useAllMaterialConsumptions } from '@/hooks/materials';
-import { ConsumptionRow } from '@/features/resources/components';
-import { ConsumptionType, consumptionTypeLabels } from '@/types/materials';
-
-const PAGE_SIZE_OPTIONS = [5, 10, 20, 50, 100];
-
-const consumptionTypeBadgeColors: Record<ConsumptionType, string> = {
-  [ConsumptionType.usedFromStock]:
-    'bg-blue-100 text-blue-700 dark:bg-blue-900/20 dark:text-blue-400',
-  [ConsumptionType.transferred]:
-    'bg-purple-100 text-purple-700 dark:bg-purple-900/20 dark:text-purple-400',
-};
+import { ConsumptionTable } from '@/features/material-consumptions/components';
+import { ConsumptionType } from '@/types/materials';
 
 export default function MaterialConsumptionsPage() {
-  const router = useRouter();
   const { data: consumptions = [], isLoading } = useAllMaterialConsumptions();
 
   const [searchQuery, setSearchQuery] = useState('');
@@ -91,8 +54,7 @@ export default function MaterialConsumptionsPage() {
 
   const totalPages = Math.ceil(filtered.length / itemsPerPage);
   const startIndex = (currentPage - 1) * itemsPerPage;
-  const endIndex = Math.min(startIndex + itemsPerPage, filtered.length);
-  const paginated = filtered.slice(startIndex, endIndex);
+  const paginated = filtered.slice(startIndex, startIndex + itemsPerPage);
 
   const now = new Date();
   const thisMonth = consumptions.filter((c) => {
@@ -130,7 +92,6 @@ export default function MaterialConsumptionsPage() {
         }
       />
 
-      {/* Stats */}
       <Card className="gap-0 p-6">
         <div className="sm:divide-border grid grid-cols-2 gap-3 sm:grid-cols-4 sm:gap-0 sm:divide-x">
           <div className="flex flex-col gap-1 rounded-lg p-3 sm:rounded-none sm:pr-6">
@@ -198,152 +159,36 @@ export default function MaterialConsumptionsPage() {
         </div>
       </Card>
 
-      {/* Table Card */}
-      <Card>
-        <CardHeader className="flex flex-row items-center gap-3 border-b px-4 py-1">
-          <div className="relative w-full max-w-xs">
-            <Search className="absolute top-1/2 left-3 size-3.5 -translate-y-1/2 text-zinc-400" />
-            <Input
-              value={searchQuery}
-              onChange={(e) => {
-                setSearchQuery(e.target.value);
-                setCurrentPage(1);
-              }}
-              placeholder="Search by material, project, task or person…"
-              className="h-8 pl-8 text-sm"
-            />
-          </div>
-          <Select
-            value={typeFilter}
-            onValueChange={(v) => {
-              setTypeFilter(v as ConsumptionType | 'all');
-              setCurrentPage(1);
-            }}
-          >
-            <SelectTrigger className="h-8 w-36 text-xs">
-              <SelectValue placeholder="Type" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">All Types</SelectItem>
-              {Object.values(ConsumptionType).map((t) => (
-                <SelectItem key={t} value={t}>
-                  {consumptionTypeLabels[t]}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <Select
-            value={projectFilter}
-            onValueChange={(v) => {
-              setProjectFilter(v);
-              setCurrentPage(1);
-            }}
-          >
-            <SelectTrigger className="h-8 w-36 text-xs">
-              <SelectValue placeholder="Project" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">All Projects</SelectItem>
-              {projectOptions.map((p) => (
-                <SelectItem key={p} value={p}>
-                  {p}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-          <div className="ml-auto flex items-center gap-2 border-l pl-3">
-            <span className="text-xs whitespace-nowrap text-zinc-500">
-              Rows per page
-            </span>
-            <Select
-              value={String(itemsPerPage)}
-              onValueChange={(v) => {
-                setItemsPerPage(Number(v));
-                setCurrentPage(1);
-              }}
-            >
-              <SelectTrigger className="h-8 w-16 text-xs">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                {PAGE_SIZE_OPTIONS.map((n) => (
-                  <SelectItem key={n} value={String(n)}>
-                    {n}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-        </CardHeader>
-
-        {paginated.length > 0 ? (
-          <>
-            <CardContent className="p-0">
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className="pl-6">Date</TableHead>
-                    <TableHead>Material</TableHead>
-                    <TableHead>Type</TableHead>
-                    <TableHead>Quantity</TableHead>
-                    <TableHead>Project</TableHead>
-                    <TableHead>Task</TableHead>
-                    <TableHead className="pr-6">Recorded By</TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {paginated.map((c) => (
-                    <ConsumptionRow
-                      key={c.id}
-                      consumption={c}
-                      onClick={() =>
-                        router.push(
-                          routes.resources.materialConsumptions.detail(c.id)
-                            .href
-                        )
-                      }
-                    />
-                  ))}
-                </TableBody>
-              </Table>
-            </CardContent>
-            <div className="flex items-center justify-between border-t px-4 py-2">
-              <span className="text-sm text-zinc-500">
-                {startIndex + 1}–{endIndex} of {filtered.length} record
-                {filtered.length === 1 ? '' : 's'}
-              </span>
-              <Pagination
-                currentPage={currentPage}
-                totalPages={totalPages}
-                onPageChange={setCurrentPage}
-              />
-            </div>
-          </>
-        ) : (
-          <CardContent>
-            <Empty variant="default">
-              <EmptyMedia variant="icon">
-                <FlameKindling className="size-6" />
-              </EmptyMedia>
-              <EmptyHeader>
-                <EmptyTitle>No consumptions found</EmptyTitle>
-                <EmptyDescription>
-                  {hasActiveFilters
-                    ? 'No records match your search. Try adjusting your filters.'
-                    : 'Record your first material consumption to get started.'}
-                </EmptyDescription>
-              </EmptyHeader>
-              {!hasActiveFilters && (
-                <Button asChild>
-                  <Link href={routes.resources.materialConsumptions.new}>
-                    Record Consumption
-                  </Link>
-                </Button>
-              )}
-            </Empty>
-          </CardContent>
-        )}
-      </Card>
+      <ConsumptionTable
+        paginated={paginated}
+        filteredCount={filtered.length}
+        startIndex={startIndex}
+        itemsPerPage={itemsPerPage}
+        onItemsPerPageChange={(n) => {
+          setItemsPerPage(n);
+          setCurrentPage(1);
+        }}
+        currentPage={currentPage}
+        totalPages={totalPages}
+        onPageChange={setCurrentPage}
+        hasActiveFilters={hasActiveFilters}
+        searchValue={searchQuery}
+        onSearchChange={(v) => {
+          setSearchQuery(v);
+          setCurrentPage(1);
+        }}
+        typeFilter={typeFilter}
+        onTypeChange={(v) => {
+          setTypeFilter(v as ConsumptionType | 'all');
+          setCurrentPage(1);
+        }}
+        projectFilter={projectFilter}
+        onProjectChange={(v) => {
+          setProjectFilter(v);
+          setCurrentPage(1);
+        }}
+        projectOptions={projectOptions}
+      />
     </div>
   );
 }
