+++ features/material-consumptions/components/index.ts
@@ -1,2 +1,4 @@
 export * from './consumption-info-card';
 export * from './consumption-context-card';
+export { ConsumptionTable } from './consumption-table';
+export { ConsumptionRow } from './consumption-row';
